Fix colormap error (#76)
* :bug: test if specifiying minimal version of matplotlib works

- colorbar need location parameter

* :zap: switch to separate environments for execution

- move tests around a bit
- comparions needs it's own environment (ToDo)

! still needs to be documented if this works

* :bug: specify environments in v2 workflow, add ipykernel as dep in action

* :bug: conda environment not allowed with run directive

* :memo: update link to miniforge

* :bug: specify pimms env and move code to script

- run does not allow conda env to be specified

* :bug: restrict matplotlib to 3.8 or lower

- looks matplotlib change their API in 3.9

* :bug: clean-up environments to free disk space

- out-of-memory in GitHub runner

* :bug: remove .snakemake folder manually

- disk space wasn't freed otherwise
- one global pimms environment next to a R environment might be the better solution.

* :art: update miniconda

- list files in project folder (to check if .snakemake is deleted)

* :white_check_mark: start testing more functionality

* :construction: fix matplotlib version for now

- seaborn lock needs to be removed to fix that.

* :construction: intermediate save

- to revert

* :bug: fixture are cached - reload explicitly

Some state fails (both test work individually with the fixtures, but not if the other is run first - the second one fails)
Can be both orders...

fix: reload the data explicitly, no caching

* :bug: pytest error due to side effects

- running order matters...
  - `test_errors.py` fails if `test_models.py`
     is run before. I don't know fastai or other might change in the state.
  - running only `test_errors.py` works fine, running it before as well.

- made some adoptions to other tests which clutter the commit a bit.

* :bug: ignore dtype for windows

- int32 vs int64 defaults on different OS?

* :bug: fix line endind error in string for windows

* :bug: another Win line break error

* :art: don't create files on the fly per default

- change folder to None and pass explicitly

---------
