+++ .github/workflows/workflow_website.yaml
@@ -16,43 +16,45 @@ jobs:
       fail-fast: false
     steps:
     - name: Checkout
-      uses: actions/checkout@v2
+      uses: actions/checkout@v4
     - name: Set up Miniconda
       # ! change action https://github.com/mamba-org/setup-micromamba
-      uses: conda-incubator/setup-miniconda@v2
+      uses: conda-incubator/setup-miniconda@v3
       with: 
-        miniforge-variant: Mambaforge
-        # miniforge-version: latest
         use-mamba: true
-        channel-priority: disabled
-        python-version: "3.8"
-        environment-file: environment.yml
-        activate-environment: pimms
+        # channel-priority: true
+        python-version: "3.10"
+        environment-file: snakemake_env.yml
+        activate-environment: snakemake
         auto-activate-base: true
-        # auto-update-conda: true
     - name: Dry-run workflow
       run: | 
         cd project
-        snakemake -s workflow/Snakefile_v2.smk --configfile config/alzheimer_study/config.yaml -p -c1 -n
+        snakemake -s workflow/Snakefile_v2.smk --configfile config/alzheimer_study/config.yaml -p -c1 --use-conda -n
     - name: Run demo workflow (integration test)
       continue-on-error: true
       run: | 
         cd project
-        snakemake -s workflow/Snakefile_v2.smk --configfile config/alzheimer_study/config.yaml -p -c4 -k
+        snakemake -s workflow/Snakefile_v2.smk --configfile config/alzheimer_study/config.yaml -p -c4 -k --use-conda
     - name: Run demo workflow again (in case of installation issues)
       continue-on-error: true
       run: | 
         cd project
-        snakemake -s workflow/Snakefile_v2.smk --configfile config/alzheimer_study/config.yaml -p -c4 -k
+        snakemake -s workflow/Snakefile_v2.smk --configfile config/alzheimer_study/config.yaml -p -c4 -k --use-conda
     - name: Run demo workflow again (in case of installation issues) with one thread
-      continue-on-error: true
       run: | 
         cd project
-        snakemake -s workflow/Snakefile_v2.smk --configfile config/alzheimer_study/config.yaml -p -c1 -k
+        snakemake -s workflow/Snakefile_v2.smk --configfile config/alzheimer_study/config.yaml -p -c1 -k --use-conda
     - name: Run differential analysis workflow
       run: | 
         cd project
-        snakemake -s workflow/Snakefile_ald_comparison.smk --configfile config/alzheimer_study/comparison.yaml -p -c4
+        snakemake -s workflow/Snakefile_ald_comparison.smk --configfile config/alzheimer_study/comparison.yaml -p -c4 --use-conda
+    - name: clean-up environments and pkg cache
+      run: |
+        conda clean -a -y
+        cd project
+        rm -r .snakemake
+        ls -la
     - name: Install website dependencies
       run: |
         pip install .[docs]
