+++ pimmslearn/plotting/errors.py
@@ -109,7 +109,7 @@ def plot_errors_by_median(pred: pd.DataFrame,
     return ax, errors
 
 
-def get_data_for_errors_by_median(errors: pd.DataFrame, feat_name, metric_name):
+def get_data_for_errors_by_median(errors: pd.DataFrame, feat_name, metric_name, seed=None):
     """Extract Bars with confidence intervals from seaborn plot.
       Confident intervals are calculated with bootstrapping (sampling the mean).
 
@@ -119,7 +119,7 @@ def get_data_for_errors_by_median(errors: pd.DataFrame, feat_name, metric_name):
 
     plotter = _BarPlotter(data=errors, x=x_axis_name, y=metric_name, hue='model',
                           order=None, hue_order=None,
-                          estimator="mean", errorbar=("ci", 95), n_boot=1000, units=None, seed=None,
+                          estimator="mean", errorbar=("ci", 95), n_boot=1000, units=None, seed=seed,
                           orient=None, color=None, palette=None, saturation=.75, width=.8,
                           errcolor=".26", errwidth=None, capsize=None, dodge=True)
     ax = plt.gca()
@@ -135,6 +135,26 @@ def get_data_for_errors_by_median(errors: pd.DataFrame, feat_name, metric_name):
     ), columns=['bin', 'model'])
     plotted = pd.concat([_index, plotted], axis=1)
     return plotted
+
+
+# def get_data_for_errors_by_median_v2(errors: pd.DataFrame, feat_name, metric_name):
+#     from seaborn._statistics import (
+#         EstimateAggregator,
+#         WeightedAggregator,
+#     )
+#     from seaborn.categorical import _CategoricalAggPlotter, WeightedAggregator, EstimateAggregator
+#     p = _CategoricalAggPlotter(
+#         data=data,
+#         variables=dict(x=x, y=y, hue=hue, units=units, weight=weights),
+#         order=order,
+#         orient=orient,
+#         color=color,
+#         legend=legend,
+#     )
+
+#     agg_cls = WeightedAggregator if "weight" in p.plot_data else EstimateAggregator
+#     aggregator = agg_cls(estimator, errorbar, n_boot=n_boot, seed=seed)
+#     err_kws = {} if err_kws is None else normalize_kwargs(err_kws, mpl.lines.Line2D)
 
 
 def plot_rolling_error(errors: pd.DataFrame, metric_name: str, window: int = 200,
