+++ project/workflow/envs/pimms.yaml
@@ -14,7 +14,7 @@ dependencies:
   - pandas>=1
   - scipy>=1.6
   # plotting
-  - matplotlib
+  - matplotlib<3.9
   - python-kaleido
   - plotly
   - seaborn<0.13
