+++ tests/plotting/expected_plotted.csv
@@ -1,0 +1,121 @@
+,bin,model,mean,ci_low,ci_high
+0,"24
+(N=12)",TRKNN,0.9571310038022816,0.5919514259208942,1.4279737379366644
+1,"24
+(N=12)",KNN,1.0388875787487997,0.6144741245086675,1.5299380106659042
+2,"24
+(N=12)",RF,1.1124649753961178,0.691806043612132,1.5553591564283435
+3,"24
+(N=12)",KNN_IMPUTE,0.9981325917398811,0.5762349029722492,1.5067300253192202
+4,"24
+(N=12)",SEQKNN,0.9050359216564058,0.4652077893514783,1.4439827634006424
+5,"25
+(N=127)",TRKNN,0.7693489972934224,0.6357217156989873,0.9497433633327523
+6,"25
+(N=127)",KNN,0.8536596417226517,0.7135306152589986,1.0250273265067762
+7,"25
+(N=127)",RF,0.8535756906362711,0.7111528884924705,1.0393605516684343
+8,"25
+(N=127)",KNN_IMPUTE,0.9241635709753512,0.7838804606928972,1.0940881823241728
+9,"25
+(N=127)",SEQKNN,0.8846409925187994,0.7503950830273517,1.0530920566813577
+10,"26
+(N=147)",TRKNN,0.5887363194720662,0.5090844094769653,0.6736474136741385
+11,"26
+(N=147)",KNN,0.604860536929101,0.5187410006420672,0.6819114805232184
+12,"26
+(N=147)",RF,0.6063602996016267,0.5291758828997412,0.6896767106811683
+13,"26
+(N=147)",KNN_IMPUTE,0.5881132943225584,0.5049176769732645,0.6801424937462395
+14,"26
+(N=147)",SEQKNN,0.6479505215712921,0.5641081191485577,0.7376662685131032
+15,"27
+(N=154)",TRKNN,0.6255527020851491,0.5357387418307774,0.7350288907938297
+16,"27
+(N=154)",KNN,0.6141361709706762,0.522468832809332,0.7176225569602876
+17,"27
+(N=154)",RF,0.6243218695141952,0.522614830543899,0.7398458074104532
+18,"27
+(N=154)",KNN_IMPUTE,0.6021011544561888,0.5115552363796041,0.7075751160756218
+19,"27
+(N=154)",SEQKNN,0.6506423432871055,0.5598981396785289,0.7563743087868408
+20,"28
+(N=147)",TRKNN,0.44208707451902934,0.3669394895815742,0.528039769361718
+21,"28
+(N=147)",KNN,0.47000548406169007,0.39011290304211876,0.5568724303785095
+22,"28
+(N=147)",RF,0.48198867625547387,0.40457173152655707,0.567013115024583
+23,"28
+(N=147)",KNN_IMPUTE,0.46617165385983894,0.38010923926175955,0.5637710195338825
+24,"28
+(N=147)",SEQKNN,0.47028125495947337,0.37857425477366496,0.5699774730414015
+25,"29
+(N=88)",TRKNN,0.37992581862386904,0.3117785433444542,0.4576554845707188
+26,"29
+(N=88)",KNN,0.4238661699756483,0.34335425677860415,0.5098219579540249
+27,"29
+(N=88)",RF,0.39455299127374305,0.3181173357362308,0.482134765962035
+28,"29
+(N=88)",KNN_IMPUTE,0.4041060161685886,0.3406706738518154,0.4771152051491808
+29,"29
+(N=88)",SEQKNN,0.4203669298141934,0.34489736016732353,0.5012959248520107
+30,"30
+(N=71)",TRKNN,0.27212864251121455,0.21852414357554373,0.330994553340812
+31,"30
+(N=71)",KNN,0.309820924888334,0.24934573323795098,0.37883563259736436
+32,"30
+(N=71)",RF,0.2960830919482679,0.24740153767324471,0.3530367460828999
+33,"30
+(N=71)",KNN_IMPUTE,0.31279045278140766,0.25650532210303933,0.3744535876738253
+34,"30
+(N=71)",SEQKNN,0.3182128352050458,0.25603345487409557,0.3894893264545288
+35,"31
+(N=61)",TRKNN,0.21178312939454516,0.17136474974252427,0.26001194347066436
+36,"31
+(N=61)",KNN,0.28925427623663624,0.21822314041240717,0.38008952442840316
+37,"31
+(N=61)",RF,0.27690123482897827,0.20873259770285557,0.3553506698768176
+38,"31
+(N=61)",KNN_IMPUTE,0.24870200843697213,0.19644005951039348,0.3028114906050088
+39,"31
+(N=61)",SEQKNN,0.25338038480442343,0.2001567073738447,0.3170852315427499
+40,"32
+(N=25)",TRKNN,0.42651743479879656,0.21485653883108224,0.7036862342616284
+41,"32
+(N=25)",KNN,0.3910882874436825,0.1991034066763848,0.6917630956908625
+42,"32
+(N=25)",RF,0.404470739593834,0.19362098355615934,0.6877041686981098
+43,"32
+(N=25)",KNN_IMPUTE,0.48948063087353305,0.28326438454200353,0.7587662536149751
+44,"32
+(N=25)",SEQKNN,0.4558611905545651,0.2398939240696641,0.7228679196211341
+45,"33
+(N=20)",TRKNN,0.20379087888487266,0.12770635582889114,0.287426869993139
+46,"33
+(N=20)",KNN,0.2534427947595642,0.1724676791124156,0.355548471583177
+47,"33
+(N=20)",RF,0.2586606779237252,0.18145672612568176,0.3407790162650329
+48,"33
+(N=20)",KNN_IMPUTE,0.32364908464759135,0.24726066625466897,0.4107746811656638
+49,"33
+(N=20)",SEQKNN,0.8348181064171756,0.39303165176309685,1.5435204959390678
+50,"34
+(N=6)",TRKNN,0.24818611159901494,0.11728239094271586,0.3957089277737785
+51,"34
+(N=6)",KNN,0.2174674745924179,0.07778358332871942,0.38537980250531406
+52,"34
+(N=6)",RF,0.32240025992021754,0.13774079341201656,0.5210252004221756
+53,"34
+(N=6)",KNN_IMPUTE,0.9300364526601529,0.4673581811964927,1.471634843658368
+54,"34
+(N=6)",SEQKNN,0.9519475780786392,0.4877553831437223,1.7451538254015493
+55,"35
+(N=1)",TRKNN,0.25086595159704217,,
+56,"35
+(N=1)",KNN,0.31500551994638926,,
+57,"35
+(N=1)",RF,0.3933540954072967,,
+58,"35
+(N=1)",KNN_IMPUTE,1.3383580136005762,,
+59,"35
+(N=1)",SEQKNN,1.8119820858075215,,
